Extract connection upgrade to a separate module
